refactor(assets/Texture): move to its own file
